fix: connect event

+++ furui-ebpf/src/close.rs
@@ -1,4 +1,4 @@
-use aya_bpf::cty::c_ushort;
+use aya_bpf::cty::c_long;
 use aya_bpf::{macros::tracepoint, programs::TracePointContext};
 
 #[tracepoint]
@@ -9,6 +9,6 @@ pub fn close(ctx: TracePointContext) -> u32 {
     }
 }
 
-unsafe fn try_close(_ctx: TracePointContext) -> Result<u32, c_ushort> {
+unsafe fn try_close(_ctx: TracePointContext) -> Result<u32, c_long> {
     Ok(0)
 }
